+++ interface.py
@@ -1,0 +1,16 @@
+from chirp_data import *
+from user_data import *
+
+class UserInterface:
+  """ The 'UserInterface' class will handle taking in all user
+      input and passing it to the appropriate data class methods
+  """
+
+  def __init__(self):
+    pass
+
+  def submit_new_chirp(self):
+    """ Takes user input and submits it to create a new chirp
+        Method arguments: n/a
+    """
+    pass
