+++ user.py
@@ -1,0 +1,12 @@
+class User:
+  """ The 'User' class defines how all user objects
+      should be constructed.
+  """
+
+  def __init__(self, username, password, id):
+    """ On init, user objects will get the following properties:
+        username - the unique username of each user profile
+        password - the password to access the user's profile
+        id - the unique ID number assigned to each user profile
+    """
+    pass
